twitter_stream/requirements.txt
requests~=2.27.1
setuptools~=56.2.0
python-dotenv~=0.19.2